requirements.txt
Django
django-tinymce>=3.6.1,<3.7.0
Pillow==9.5.0
django-colorfield
django-jazzmin
openpyxl
beautifulsoup4
whitenoise
